TICKET-152: update all import paths after directory reorganization
