+++ demos/arena/src/nodes/player/LocalPlayerNode.test.ts
